refactoring component data-table. add axios requests
