+++ client/package.json
@@ -31,6 +31,7 @@
         "@tabler/icons-react": "^3.31.0",
         "@tailwindcss/postcss": "^4.0.15",
         "@tanstack/react-table": "^8.21.2",
+        "axios": "^1.8.4",
         "class-variance-authority": "^0.7.1",
         "clsx": "^2.1.1",
         "lucide-react": "^0.483.0",
